test version for Vega

+++ plugins/SuperIOSensors/Fintek718x/F718x.cpp
@@ -295,7 +295,7 @@ bool F718x::start(IOService * provider)
 			name = OSDynamicCast(OSString, configuration->getObject(key));
 		}
 		
-		UInt32 nameLength = name ? strlen(name->getCStringNoCopy()) : 0;
+		size_t nameLength = name ? strlen(name->getCStringNoCopy()) : 0;
 		
 		if (readTachometer(i) > 10 || nameLength > 0)
 			if (!addTachometer(i, (nameLength > 0 ? name->getCStringNoCopy() : 0)))
